% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAPIStatus.R
\name{GetAPIStatus}
\alias{GetAPIStatus}
\title{Returns the API status}
\usage{
GetAPIStatus()
}
\value{
A message containing the API status
}
\description{
Some placeholder
}
\examples{
\dontrun{
GetAPIStatus()
}

}
