% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAPIVersion.R
\name{GetAPIVersion}
\alias{GetAPIVersion}
\title{GetAPIVersion}
\usage{
GetAPIVersion()
}
\value{
A character string containing the API version
}
\description{
Returns the API verion
}
\examples{
\dontrun{
GetAPIVersion()
}

}
