% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMarketTime.R
\name{GetMarketTime}
\alias{GetMarketTime}
\title{Returns the current market time}
\usage{
GetMarketTime()
}
\value{
A tibble containing the symbol, with last, bid, and ask prices.
}
\description{
Some placeholder
}
\examples{
\dontrun{
GetQuotes(APIKeys, "XYZ")
}

}
