% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOptionDates.R
\name{GetOptionDates}
\alias{GetOptionDates}
\title{Returns expiration dates on a single underlying}
\usage{
GetOptionDates(APIKeys, symbol)
}
\arguments{
\item{APIKeys}{A vector containing API keys (order: cKey, cSecret, oKey, oSecret).}

\item{symbol}{The underlying's symbol}
}
\value{
A tibble containing available option expiration dates and days to expiration
}
\description{
Some placeholder
}
\examples{
\dontrun{
GetOptionDates(APIKeys, "XYZ"
}

}
