% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetQuotes.R
\name{GetQuote}
\alias{GetQuote}
\title{GetQuote}
\usage{
GetQuote(APIKeys, symbol)
}
\arguments{
\item{APIKeys}{A vector containing API keys (order: cKey, cSecret, oKey, oSecret).}

\item{symbol}{The underlying's symbol}
}
\value{
A tibble containing the symbol, with last, bid, and ask prices.
}
\description{
Returns market quote (last, bid, ask) on a single stock. Requires API keys (in specific order). See help for more info.
}
\examples{
\dontrun{
GetQuotes(APIKeys, "XYZ")
}

}
