% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAccountHoldings.R
\name{GetAccountHoldings}
\alias{GetAccountHoldings}
\title{GetAccountHoldings}
\usage{
GetAccountHoldings(APIKeys, acctNum)
}
\arguments{
\item{APIKeys}{A vector containing API keys (order: cKey, cSecret, oKey, oSecret).}

\item{acctNum}{An Ally Invest account number.}
}
\value{
A tibble containing curret account holdings for the account passed..
}
\description{
Returns information on your current account holdings.  Requires API keys (in specific order). See help for more info.
}
\examples{
\dontrun{
GetAccountHoldings (APIKeys, xxxxxxxx)
}


}
