% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMarketTime.R
\name{GetMarketTime}
\alias{GetMarketTime}
\title{GetMarketTime}
\usage{
GetMarketTime()
}
\value{
A tibble containing the symbol, with last, bid, and ask prices.
}
\description{
Returns the current market time.
}
\examples{
\dontrun{
GetMarketTime()
}

}
