% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNewsHeadlines.R
\name{GetNewsHeadlines}
\alias{GetNewsHeadlines}
\title{GetNewsHeadlines}
\usage{
GetNewsHeadlines(APIKeys, symbol, maxhits, enddate)
}
\arguments{
\item{symbol}{A ticker}

\item{maxhits}{The number of hits to display}

\item{authKeys}{A vector containing API keys (order: cKey, cSecret, oKey, oSecret).}

\item{endDate}{The end date, passed as two-digit month and two digit year, separated by a '/' (mm/dd)}
}
\value{
A tibble containing news headlines from todays date to \code{enddate}
}
\description{
Returns news articles based on symbol and dates
}
\examples{
\dontrun{
GetNewsHeadlines(APIKeys, "aapl", "10", "02/27")
}

}
