% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOrders.R
\name{GetOrders}
\alias{GetOrders}
\title{GetOrders}
\usage{
GetOrders(APIKeys, acctNum)
}
\arguments{
\item{APIKeys}{A character vector containing API keys (order: cKey, cSecret, oKey, oSecret).}

\item{acctNum}{Your Ally Invest account number}
}
\value{
A tibble containing information on recent orders
}
\description{
Returns recent orders. Requires API keys (in specific order). See help for more info.
}
\examples{
\dontrun{
GetOrders(APIKeys, xxxxxxxx)
}

}
